+++ src/cli/domain/package-server-script/bundle/index.js
@@ -1,15 +1,15 @@
 /*jshint camelcase:false */
 'use strict';
-
+var _ = require('underscore');
+var config = require('./webpack.server.config');
 var console = require('console');
-var webpack = require('webpack');
-var path = require('path');
 var MemoryFS = require('memory-fs');
+var path = require('path');
+var webpack = require('webpack');
 
 var memoryFs = new MemoryFS();
 
 function bundle(dataPath, fileName, options, callBack) {
-
   if (typeof options === 'function') {
     callBack = options;
     options = {
@@ -22,50 +22,21 @@ function bundle(dataPath, fileName, options, callBack) {
     };
   }
 
-
-  var webpackConfig = {
-    entry: dataPath,
-    output: {
-      path: '/build',
-      filename: fileName
+  var webpackConfig = _.extend(
+    {
+      entry: dataPath,
+      output: {
+        path: '/build',
+        filename: fileName
+      }
     },
-    plugins: [
-      new webpack.optimize.OccurenceOrderPlugin(),
-      new webpack.optimize.UglifyJsPlugin({
-        compressor: {
-          warnings: false,
-          screw_ie8: true
-        }
-      }),
-      new webpack.DefinePlugin({
-        'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV)
-      })
-    ],
-    module: {
-      loaders: [
-        {
-          test: /\.jsx?$/,
-          exclude: /node_modules/,
-          loader: 'babel',
-          query: {
-            'presets': [
-              ['env', {
-                'targets': {
-                  'node': 'current'
-                }
-              }]
-            ]
-          }
-        }
-      ]
-    }
-  };
+    config
+  );
 
   var compiler = webpack(webpackConfig);
   compiler.outputFileSystem = memoryFs;
 
   compiler.run(function(err, stats){
-
     // handleFatalError
     var error = err;
     if (error) {
@@ -73,7 +44,6 @@ function bundle(dataPath, fileName, options, callBack) {
     }
 
     var info = stats.toJson();
-
     // handleSoftErrors
     if (stats.hasErrors()) {
       error = info.errors;
