+++ src/cli/domain/package-server-script/bundle/webpack.server.config.js
@@ -1,0 +1,44 @@
+/*jshint camelcase:false */
+'use strict';
+var webpack = require('webpack');
+
+var config = {
+  module: {
+    loaders: [
+      {
+        test: /\.json$/,
+        loader: 'json-loader',
+        exclude: /node_modules/
+      },
+      {
+        test: /\.jsx?$/,
+        exclude: /node_modules/,
+        loader: 'babel-loader',
+        query: {
+          'presets': [
+            ['env', {
+              'targets': {
+                'node': 0.10
+              }
+            }]
+          ]
+        }
+      }
+    ]
+  },
+  plugins: [
+    new webpack.optimize.OccurenceOrderPlugin(),
+    new webpack.optimize.UglifyJsPlugin({
+      compressor: {
+        warnings: false,
+        screw_ie8: true
+      },
+      sourceMap: false
+    }),
+    new webpack.DefinePlugin({
+      'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV)
+    })
+  ]
+};
+
+module.exports = config;
